fix: migrate image-proxy to Hono and delete dead App Router files
