--- langwatch/src/app/image-proxy/route.ts
@@ -1,41 +1,0 @@
-import { type NextRequest, NextResponse } from "next/server";
-import { ssrfSafeFetch } from "../../utils/ssrfProtection";
-
-export async function GET(req: NextRequest) {
-  const url = req.nextUrl.searchParams.get("url");
-  if (!url) {
-    return NextResponse.json({ error: "Missing url" }, { status: 400 });
-  }
-
-  try {
-    const response = await ssrfSafeFetch(url);
-
-    if (!response.ok) {
-      return NextResponse.json(
-        { error: `Failed to fetch image: ${response.statusText}` },
-        { status: response.status },
-      );
-    }
-
-    const contentType = response.headers.get("content-type");
-    if (!contentType?.startsWith("image/")) {
-      return NextResponse.json(
-        { error: "URL does not point to an image" },
-        { status: 400 },
-      );
-    }
-
-    const imageBuffer = await response.arrayBuffer();
-    return new NextResponse(imageBuffer, {
-      headers: {
-        "Content-Type": contentType,
-        "Cache-Control": "public, max-age=31536000",
-      },
-    });
-  } catch {
-    return NextResponse.json(
-      { error: "Failed to fetch image" },
-      { status: 500 },
-    );
-  }
-}
